Refactor code for custom RPC and migration
Made revisions to the custom RPC URL and network arguments in 'getAssetsForMigration.ts' and 'migrationsClient.ts'. Instead of using a simple conditional operator, a custom object is added when the network is a custom fork. This helps to improve function parameters management and make it easier to handle potential future changes.