Refactor server rewrites to accept function API URL as parameter
The server rewrite rules are updated to use a provided "functionApiUrl" parameter instead of directly accessing the environment variable. This is achieved by getting the "functionApiUrl" from the request cookies. If no such cookie exists, it falls back to using the environment variable. This allows for more flexibility when setting the functions API URL.