Add PayPal USD token and Steakhouse Vaults (#3685)
The token 'PayPal USD' (PYUSD) has been added to the blockchain's token metadata list. Additionally, lending protocols have been updated with details for the Stakehouse on different assets. The list of protocols on the product hub has also been updated to include 'erc-4626'. This change reflects our continuing effort to include a variety of financial options for our users on the platform.