Add new token metadata and update libraries (#3716)
Expanded the token-metadata-list to include additional token configurations. Updated the token pull list in the `erc-4626` protocol settings. Updated the versions of `@oasisdex/addresses` and `@oasisdex/dma-library` in `package.json` to match the respective updates in `yarn.lock`. Applied minor changes to the deposit form. The additional tokens support more options for users in token transactions, enhancing the overall functionality of the platform.