chore: [sc-15087] Product Hub - Issue with DAI filtering (#3763)
* Remove specific pools from rewards list

This commit removes 'DEGEN-CUSDCV3' and 'SNX-CUSDCV3' from the list of pools that have rewards. This change reflects the current state of rewards distribution and ensures only valid pools are included in the list.

* Update token group conditions in ajna handler

This commit modifies the token group conditions in ajnaHandler.ts. Changes were made to the conditions that determine whether to set primaryTokenGroup and secondaryTokenGroup, ensuring the inclusion of valid tokens only, thereby maintaining the accuracy of data in the product hub.