Update disconnect wallet button text and functionality (#4004)
The disconnect wallet button text and functionality have been updated to provide a clearer user experience. The button now displays different text and color based on whether the user is holding the Shift key or not. If the user is holding the Shift key, the button text is changed to "Clear all data and Disconnect Wallet" and the button color is set to a warning color. Otherwise, the button text remains as "Disconnect Wallet" and the color is set to the primary color. Additionally, the disconnect callback now accepts a boolean parameter to determine whether to clear the local storage or not. This change improves the usability of the user settings page.